Change layout; add color, CIDR tracking
This changes the layout of the graph. Instead of allowing dot to generate a
mess of curves, nodes are positioned diagonally with orthogonal edges. The
graph can now be rendered with neato. Subgraphs are not displayed due to this
change.

Also, this adds color to the graph. Security groups are shown in one of several
colors to distinguish them. Text for managed security groups is black. Edges
are the color of the security group they connect to, with labels to match.

Also, this adds explicit CIDR graph nodes. Previously, the graph nodes
for CIDRs implicitly existed due to their edges. CIDRs are now tracked and
included in the layout. They are octagonal.